+++ modules/role/model.go
@@ -3,7 +3,7 @@ package role
 import "time"
 
 type Role struct {
-	ID          uint       `gorm:"primary_key" gorm:"column:_id" json:"_id"`
+	ID          uint64     `gorm:"primary_key" gorm:"column:_id" json:"_id"`
 	Name        string     `gorm:"column:name" json:"name"`
 	Description string     `gorm:"column:description" json:"description"`
 	Config      string     `gorm:"column:config" json:"config" sql:"json"`
